requirements.txt
psycopg2
python-dotenv
boto3
botocore
fpdf
Pillow
requests
yfinance
matplotlib
pdf2image